Update WP
I changed an Acronym to try and fix the "crap" acronym. 
Made it MONKEY CRYTPO ASSET EXCHANGE PLATFORM (MCAXP).

Also, as the Blockchain Transfer Gateways sounded confusing for fungible vs. conversion I called bidirectional transfers via BTGs and conversion to MIX a Token Immigration Gateway (TIG).  

Question -- what if someone transfers PRE to the COA blockchain, what does that do to POV mining? and the rights of PRE that was transferred to the COA blockchain with regard to MIX mining reward?

Question -- what are examples of utility for blockchain transfers -- what is utility of COA on PRE blockchain?